Add timing to test + bug fixes
* New test file
* Fixed bug with wrong index for falsE check
* Make '.' check more clear

+++ C++/jwcfg/Utilities.cpp
@@ -60,7 +60,7 @@ int FindFirstOf(const c8* str, u32 strLen, c8 ch)
 	for(; idx < (int)strLen; ++idx)
 	{
 		if (str[idx] == ch)
-			return (int)idx;
+			return idx;
 	}
 
 	return -1;
